[Feat] 아이템 사용 후 체력 랜덤 증감 구현
TODO
1. 스피드 증감 구현 필요
2. 아이템 위치 조정 필요
3. 아이템 효과 표시해줄 텍스트 구현 필요

+++ Assets/Scripts/1. Player/PlayerCondition.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public interface IDamagable
 {
-    void TakePhysicalDamage(int damage);
+    void TakePhysicalDamage(float damage);
 }
 
 
@@ -39,7 +39,7 @@ public class PlayerCondition : MonoBehaviour, IDamagable
         // GameOverUI 생성 후 restart할 수 있게.
     }
 
-    public void TakePhysicalDamage(int damage)
+    public void TakePhysicalDamage(float damage)
     {
         health.Subtract(damage);
         onTakeDamage?.Invoke();
